Few wrappers done.
I got a two renders seemmingly correct on the REPL, but I have a weird
syntax bug on the case classes, it's honestely just eerie, it happened
like 4 times today: I write a class in the main src tree and it does not
compile. I write it in the REPL, and it compiles, and I write it to a
single file outside the main src and it does compile, and I copy-paste
it to the main src and it does compile until I touch a space of
something and suddenly it dosnt't work anymore and I need to restart the
cycle.

Perhaps the real bug is that I need to go to bed.