Compiles, runs, gives us a nice .png
Single-threaded right-now, for the simple reason that there's only one
actor. The exact spreading of the work remains to be done, but it
shouldn't be all that hard; important parts are wrapped up, and that's
cool.

I made one bug, though.

I need a full, deep, pass-by-*actual*-value-and-not-value-of-a-pointer
copy constructor because of course I can't pass around shallow
references to heap objects around on a fucking *network*; I wrote one,
tested it and it doesn't work. I tested the thing with a simple
reference and it does, for the single-machine setup.

So I'l need to figure this one out before I can go on with the funky
remote actors and stuff.