YES
Changed everything after I realised the design for Master was
brain-damaged with the nested "receive" thing. All I needed was to store
the address of the original sender (the main program making the request
to the master actor) and send it a message after the master/slave dialog
is done, which goes one as a ping-pong of receives/replies.

Still only one actor, but trust me (I'm the Doctor) I know how to scale
this shit to many ones.

First, I'l pass an array of slaves to the master instead of one. The
master will send an initial message to all of them, and from now on the
ping-pong starts.

The termination is weird, though.

Right now, I have a neat little class that implements a factory for the
pixel asks, named BuildAskPixel, and it's method next; next return an
ask if there's one to do, or nothing if there isn't anymore when the
list is done. I'l try to change things up so we can use streams, because
Scala is a dumb eager language and storing like a million pixels request
in memory is somewhat dumb. I'l also make is so that it keeps track of
the the number of pixels currently rendering, perhaps by routing the
gets trough it too with the asks, or something. It'l return yet another
flag when there are no pixels left to send and no jobs running;
program's over.

It's not the most beautiful code I've written that's for sure, but damn,
actors are like another universe.

I think we're all gonna be a bit in the dark copying MapReduce or
monkeying around with them until some guy comes up with the Actor
Programming Design Patterns bible and we'l all be illuminated like in
the 90's when nobody knew what to do with their C++ classes.

Anyways.